create a O2M relationship with User and Location
it took me a while to understand how to use the edge to and from but it
does make sense once it clicked.

firstly, we need to have at least a to and from edge (to in one schema
and from in another schema) so we can create a relationship which makes
sense. in order to query the edge, we need to have a name so that's why
we need to use the Ref method and have the same name from both schemas
in order to connect it.

when we have things like required or unique, it will apply to the schema
that it lives in. so in the code here, the user is required to have a
location tied to their profile but the location node isn't required to
have users in it. we also use the unique here so we can enforce the O2M
relationship. if, at runtime (generation time), we see that there is
another location that is referencing the same user, we will throw an
error.

it is also important to know that for each node in the graph, it
represents one table in the database. we are representing each table in
the graph explicitly with the ent framework, defining its fields, and
the ent framework will create a bunch of helper methods for us to help
query, mutate, delete, etc...